transaksi booking

+++ database/migrations/2023_07_17_084619_create_booking_lists_table.php
@@ -15,8 +15,10 @@ return new class extends Migration
     {
         Schema::create('booking_lists', function (Blueprint $table) {
             $table->id();
-            $table->integer('room_id');
+            $table->unsignedBigInteger('room_id');
             $table->unsignedBigInteger('user_id');
+            // $table->integer('room_id');
+            // $table->integer('user_id');
             $table->date('date');
             $table->time('start_time');
             $table->time('end_time');
